Moved logic for embed building to separate service
