+++ DiscordBot/Features/Birthdays/IUserBirthdayEmbedBuilder.cs
@@ -1,0 +1,10 @@
+﻿using Discord;
+
+namespace DevSubmarine.DiscordBot.Birthdays
+{
+    public interface IUserBirthdayEmbedBuilder
+    {
+        Task<Embed> BuildUserBirthdayEmbedAsync(UserBirthday birthday, ulong? guildID, CancellationToken cancellationToken = default);
+        Task<Embed> BuildUpcomingBirthdaysEmbedAsync(IEnumerable<UserBirthday> birthdays, CancellationToken cancellationToken = default);
+    }
+}
