+++ DiscordBot/Features/Birthdays/UserBirthdaysDependencyInjectionExtensions.cs
@@ -13,6 +13,7 @@ namespace Microsoft.Extensions.DependencyInjection
 
             services.TryAddTransient<IUserBirthdaysStore, MongoUserBirthdayStore>();
             services.TryAddSingleton<IUserBirthdaysProvider, UserBirthdaysProvider>();
+            services.TryAddTransient<IUserBirthdayEmbedBuilder, UserBirthdayEmbedBuilder>();
 
             return services;
         }
